+++ Stepper_motor_multiple_functions/.vscode/settings.json
@@ -1,0 +1,8 @@
+{
+    "files.associations": {
+        "string.h": "c",
+        "io.h": "c",
+        "iom328p.h": "c"
+    },
+    "cmake.sourceDirectory": "${workspaceFolder}/.pio/libdeps/nanoatmega328/Unity"
+}
