+++ Stepper_motor_multiple_functions/include/usart.h
@@ -1,0 +1,12 @@
+#ifndef USART_H_INCLUDED
+#define USART_H_INCLUDED
+
+int uart_putchar(char c, FILE *stream);
+int uart_getchar(FILE *stream);
+
+void uart_init(void);
+void io_redirect(void);
+
+
+#endif
+
