+++ Stepper_motor_multiple_functions/platformio.ini
@@ -1,0 +1,20 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:nanoatmega328]
+platform = atmelavr
+board = nanoatmega328
+
+upload_port = COM8
+ 
+
+build_flags = -funsigned-char -Wl,-u,vfprintf -lprintf_flt -lm, -Wl,-u,vfscanf -lscanf_flt -lm
+
+monitor_echo = yes
