--- src/usart.c
@@ -1,60 +1,0 @@
-#include <avr/io.h>
-#include <stdio.h>
-#include "usart.h"
-
-#ifndef F_CPU
-#define F_CPU 16000000UL
-#endif
-
-#ifndef BAUD
-#define BAUD 9600
-#endif
-#include <util/setbaud.h>
-
-
-FILE f_uart = FDEV_SETUP_STREAM(uart_putchar, uart_getchar, _FDEV_SETUP_RW);
-
-
-void uart_init(void) {
-	
-	UBRR0H = UBRRH_VALUE;
-	UBRR0L = UBRRL_VALUE;
-	
-	#if USE_2X
-	UCSR0A |= _BV(U2X0);
-	#else
-	UCSR0A &= ~(_BV(U2X0));
-	#endif
-
-	UCSR0C = _BV(UCSZ01) | _BV(UCSZ00); /* 8-bit data */
-	UCSR0B = _BV(RXEN0) | _BV(TXEN0);   /* Enable RX and TX */
-	
-}
-
-void io_redirect(void ){
-		
-		
-		stdout = &f_uart;	//redirect standard output to uart
-		stdin  = &f_uart;  //redirect standard input to uart
-}
-	
-
-int uart_putchar(char c, FILE *stream) {
-	if (c == '\n') {
-		uart_putchar('\r', stream);
-	}
-	loop_until_bit_is_set(UCSR0A, UDRE0);
-	UDR0 = c;
-	return 0;
-}
-
-int uart_getchar(FILE *stream) {
-	loop_until_bit_is_set(UCSR0A, RXC0);
-	if (UCSR0A & _BV(FE0))
-		return _FDEV_EOF;
-	if (UCSR0A & _BV(DOR0))
-		return _FDEV_ERR;
-	// now USART has data available in buffer
-	
-	return UDR0;
-}
